move primary key set methods to under feats
updating workflow docs
