+++ docs/design.rst
@@ -1,0 +1,2 @@
+Design
+======
