+++ src/finagg/sec/sql.py
@@ -273,46 +273,6 @@ def get_ticker(cik: str, /, *, engine: None | Engine = None) -> str:
     return str(ticker)
 
 
-def get_ticker_set(
-    lb: int = 1,
-    *,
-    engine: None | Engine = None,
-) -> set[str]:
-    """Get all unique ticker symbols in the raw SQL tables that have at least
-    ``lb`` rows.
-
-    This method is convenient for accessing the tickers that have raw SQL data
-    associated with them so the data associated with those tickers can be
-    further refined. A common pattern is to use this method and other
-    ``get_ticker_set`` methods (such as those found in :mod:`finagg.sec.feat`)
-    to determine which tickers are missing data from other tables or features.
-
-    Args:
-        lb: Lower bound number of rows that a company must have for its ticker
-            to be included in the set returned by this method.
-        engine: Feature store database engine. Defaults to the engine
-            at :data:`finagg.backend.engine`.
-
-    Examples:
-        >>> "AAPL" in finagg.sec.sql.get_ticker_set()
-        True
-
-    """
-    engine = engine or backend.engine
-    with engine.begin() as conn:
-        tickers = (
-            conn.execute(
-                sa.select(submissions.c.ticker)
-                .join(tags, tags.c.cik == submissions.c.cik)
-                .group_by(tags.c.cik)
-                .having(sa.func.count(tags.c.filed) >= lb)
-            )
-            .scalars()
-            .all()
-        )
-    return set(tickers)
-
-
 def get_tickers_in_industry(
     *,
     ticker: None | str = None,
