+++ tests/test_sec/test_feat.py
@@ -16,7 +16,7 @@ def engine() -> Engine:
 def test_quarterly_candidate_ticker_set(engine: Engine) -> None:
     finagg.sec.feat.submissions.install({"AAPL"}, engine=engine)
     finagg.sec.feat.tags.install({"AAPL"}, engine=engine)
-    assert "AAPL" in finagg.sec.sql.get_ticker_set(engine=engine)
+    assert "AAPL" in finagg.sec.feat.tags.get_ticker_set(engine=engine)
     assert "AAPL" in finagg.sec.feat.quarterly.get_candidate_ticker_set(engine=engine)
