+++ tests/test_yfinance/test_feat.py
@@ -15,7 +15,7 @@ def engine() -> Engine:
 
 def test_daily_candidate_ticker_set(engine: Engine) -> None:
     finagg.yfinance.feat.prices.install({"AAPL"}, engine=engine)
-    assert "AAPL" in finagg.yfinance.sql.get_ticker_set(engine=engine)
+    assert "AAPL" in finagg.yfinance.feat.prices.get_ticker_set(engine=engine)
     assert "AAPL" in finagg.yfinance.feat.daily.get_candidate_ticker_set(engine=engine)
